requirements.txt
tensorflow
opencv-python
streamlit
numpy
matplotlib
scikit-learn
pandas
pillow